chore: updated header component
